Actually starting to finish refactoring.
